requirements.txt
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.1
evervault==2.0.0
evervault_attestation_bindings==0.1.0a5
Flask==2.2.3
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.17
python-dotenv==0.21.1
requests==2.28.2
urllib3==1.26.14
Werkzeug==2.2.3
zipp==3.13.0
